# server/src/docs/swagger.yml
swagger: '2.0'
info:
  version: 1.0.0
  title: TaoScan APIs
  description: TaoScan APIs
  license:
    name: Github
    url: 'https://github.com/taoblockchain/taoscan'
...
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /api/tokens:
    get:
      tags:
        - Token
      summary: Get list tokens
      parameters:
        - name: page
          in: query
          required: false
          description: default = 1
          type: number
        - name: limit
          in: query
          required: false
          description: default 20, maximum = 50
          type: number
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        500:
          description: Server Internal Error
  /api/tokens/{hash}:
    get:
      tags:
        - Tokens
      summary: Get token detail
      parameters:
        - name: hash
          in: path
          required: true
          description: token address
          type: string
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        404:
          description: Not Found
        500:
          description: Server Internal Error
  /api/tokens/{token}/holder/{holder}:
    get:
      tags:
        - Tokens
      summary: Get token holder detail
      parameters:
        - name: token
          in: path
          required: true
          description: token address
          type: string
        - name: holder
          in: path
          required: true
          description: holder address
          type: string
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        404:
          description: Not Found
        500:
          description: Server Internal Error
...